rdp: Add support for FreeRDP 2.3
FreeRDP 2.3 broke the API, which means that compiling
gnome-remote-desktop with FreeRDP 2.3 results in an error.
Additionally, some warnings are now appearing due to the API breaking
changes.

However, bumping the FreeRDP version to 2.3 is too late for Gnome 40,
so just ensure that building with FreeRDP 2.3.x just works by adding
a #ifdef around the API changes.