clipboard-rdp: Don't leak the MimeTypeTables
When the RDP client sends a new FormatList, but the session already
stopped, then gnome-remote-desktop drops the FormatList.
However, gnome-remote-desktop already prepared the MimeTypeTables for
the update. These are currently leaked.

So when the session already stopped, but gnome-remote-desktop still got
a new FormatList, free the MimeTypeTables to prevent a memory leak.