rdp-fuse-clipboard: Handle all file lists via clipDataIds
Currently, the whole file list of a uri-list is located under the FUSE
root directory.
This has the disadvantage that the FUSE implementation is only able to
handle one uri-list at the same time.
For clipboard data locking, it needs however be able to handle multiple
uri-lists at the same time.

To achieve this goal, the FUSE implementation needs to be restructured:
Instead of having the whole selection under the root directory, add
another layer: The ClipDataEntry layer.
This layer will basically contain all directories for clipDataIds, i.e.
the structure will now look like this: FUSE_ROOT/clipDataId/selection.

When clipboard data locking is not used, the FUSE implementation will
still use the same pattern:
The selection for a uri-list without clipDataId will have its own
directory.
This makes the whole implementation easier, as the code paths for
locking usage and no-locking usage are with this approach similar.

For this new handling, add new API methods for the selection handling:
- clip_data_id_new()
- clip_data_id_free()
- dismiss_all_no_cdi_requests()
- clear_no_cdi_selection()
- lazily_clear_all_cdi_selections()
- set_cdi_selection()
- set_no_cdi_selection()

A cdi selection is a selection with clipDataId, while a no-cdi
selection is a selection without clipDataId.

The clipboard behaviour for selections without clipDataId will not
change.
In the next commit, however, clipboard-rdp will be ported to the new
API.
clear_selection() then becomes clear_no_cdi_selection(),
dismiss_all_requests() becomes dismiss_all_no_cdi_requests() and
set_selection() becomes set_no_cdi_selection().

For selections with a clipDataId, g-r-d will only emit a lock, when the
client advertised file content and an application on the server side
requests the uri-list.
In that case, clipboard-rdp asks the FUSE clipboard for a new
clipDataId. If the FormatDataRequest fails, this clipDataId will then
be freed using clip_data_id_free().
Since gnome-remote-desktop caches FormatData (and therefore the
uri-list too), it will only emit the lock request per uri-list once.
If the uri-list is retrieved successfully, clipboard-rdp will then use
that id to set the selection using set_cdi_selection().

The clipDataId and the associated ClipDataEntry will live as long as
the clipboard selection doesn't change.
Upon receiving a new FormatList (from either peers), g-r-d will,
however, not directly clear the selection.
Instead, for selections with clipDataId, a timeout is set.
The reason for this is that g-r-d does not know if nautilus finished
its file copy operations or whether it aborted them.
When a file operations happens, i.e. a read() request happens or g-r-d
receives a FileContentsResponse, the timeout might be reset.
The source function to drop a ClipDataEntry (and its selection) runs
on the main thread (same as setting and clearing a (no-cdi) selection).
The read() requests and the FileContentsReponses, however, run on the
FUSE thread and the cliprdr thread.
To avoid race conditions, the timeout reset must happen on the main
thread, since glib does not provide a way to (easily) reset a timeout.
The GSource for the timeout needs to be destroyed and then recreated.
Do this with a persistent GSource: Instead of directly resetting the
timeout directly, queue the clipDataId to a HashTable and set the
timeout_reset_source to ready.
The timeout_reset_source runs on the main thread too and looks up the
ClipDataEntry of the clipDataId and resets the timeout, if the entry
still exists.

To avoid high cpu usage on the main thread, while a copy operation
happens, don't always queue a reset for the timeout.
Instead, save the time, when the (re)set of the timeout happens and
only queue a timeout, if a certain time (10 seconds) passed since the
last (re)set.
This creates an effective timeout for a ClipDataEntry of 50-60 seconds,
since the reset timeout is 60 seconds.

When a new FormatList is received (from either peer),
lazily_clear_all_cdi_selections() is called.
This method applies the initial timeout to all selections, that do not
have a timeout set yet.

While the new API is implemented in this commit, the old API, that
clipboard-rdp uses, is still usable.
Both APIs should however not be used at the same time.

While for server-to-client transfer mstsc only works with clipboard
data locking, when the (shared-)clipboard owner doesn't change, the
same applies to client-to-server transfer:
When, during a client-to-server transfer, the clipboard owner changes,
mstsc will answer the FileContentsRequests with the CB_RESPONSE_FAIL
flag, which effectively aborts the transfer (same behaviour as when not
using clipboard data locking with mstsc).
gnome-remote-desktop cannot prevent this behaviour.
mstsc has the same behaviour, when connecting to MS Windows RDS.
With FreeRDP based clients, this can be avoided in the future, as soon
as one (FreeRDP-) client implements clipboard data locking.