clipboard-vnc: Also advertise the text/plain;charset=utf-8 mime-type
When copying text on the remote desktop client side, then
gnome-remote-desktop converts that text into a UTF-8 string and
advertises it with the mime-type UTF8_STRING.

GTK4 apps, however, ignore that mime-type and instead prefer the mime-
type text/plain;charset=utf-8.

So, advertise the UTF-8 string with the mime-type
text/plain;charset=utf-8 too, to be able to paste text into GTK4
applications on the server side too.