clipboard-rdp: Also advertise the text/plain;charset=utf-8 mime-type
When copying text on the remote desktop client side, then
gnome-remote-desktop advertises the mime-type UTF8_STRING and converts
the text into a UTF-8 string, when an application requests the text.

GTK4 apps, however, ignore that mime-type and instead prefer the mime-
type text/plain;charset=utf-8.

So, advertise the mime-type text/plain;charset=utf-8 too, to be able to
paste text into GTK4 applications on the server side too.
The content of both mime-types is the same.