rdp: Add RDP surface
This is a preparatory step for the graphics pipeline.
While in the legacy path frame updates happen directly in the graphics
output buffer, that is visible to the user, the graphics pipeline
handles frame updates differently:

Instead of updating one chunk of area in the graphics output buffer,
the graphics pipeline only updates an offscreen surface.
That surface can be mapped to the user-visible graphics output
buffer (to become an onscreen surface) for the usage as a monitor or
window (RAIL), but is not limited to that.
The graphics pipeline also allows using offscreen surfaces to e.g.
composite frames with the usage of the blitting PDUs or to cache frame
content using them.

Each RDP surface can later correspond to a GFX surface, but does not
have to, to remain compatible with the legacy path.
In the next step, gnome-remote-desktop will be adapted to use RDP
surfaces, when handling frame updates.