rdp: Adapt to GrdRdpSurface
With the introduction of the RDP surface in the last commit, adapt the
frame handling in session-rdp to it.
RDP surfaces can be invalidated in case of e.g. a recreation is
necessary.
This will for example be the case, when the surface was resized, as
GFX surfaces cannot be directly resized.
Instead, GFX surfaces need to be recreated. This will later also
force the codec to reset, in case of the frame is progressively
encoded (e.g. when using H264 or RFX Progressive with TILE_FIRST +
TILE_UPGRADE tiles).