session-rdp: Set RLGR mode before encoding frame data
RFX and RFX Progressive are similar codecs. However, they are not the
same and even when progressive encoding is not used, the encoded frame
will not be the same.
RFX uses for the Golomb-Rice coding the RLGR3 mode, while RFX
Progressive uses the RLGR1 mode.

Since, with the introduction of the graphics pipeline, both are
supported, set the RLGR mode before encoding the frame data to ensure
that the correct mode is used.