rdp: Add GSource to encode pending frame data
Add a GSource to encode the pending frame data. This GSource will then
be later used to update all RDP surfaces to their latest frame.

To be able to do that, also save the pending frame for an RDP surface,
when the current situation does not allow updating an RDP surface, but
a later situation will.