session-rdp: Encode pending frame data upon end of SuppressOutput
Use the GSource, that was added in the previous commit, to encode any
pending frame data, when client stops suppressing the output.
This will be the case, when the user restores the RDP client or
switches to the workspace, where the restored RDP client window lies.
Without this commit, the user would have to trigger a new frame by e.g.
moving the mouse to ensure that they received the latest frame content.