rdp-server: Run primitives benchmark when the server starts
For different actions, like colour conversion, FreeRDP uses the FreeRDP
primitives.
When using the FreeRDP primitives the first time, FreeRDP runs a small
benchmark.
This benchmark can take up to ~310ms. Running the benchmark earlier
(, when the server starts), saves that time, when the first RDP client
connects.

So, add a primitives_get() call, when the RDP server initializes to
ensure that the benchmark won't have to run any more, when the user
connects.