session-rdp: Add status flags for the graphics pipeline
The graphics pipeline is a dynamic channel and has its own capability
exchange.
This means: If the connection is activated, then the graphics pipeline
will not be ready yet.
The RDP_PEER_PENDING_GFX_INIT flag signals, that there is a pending
capability exchange (graphics pipeline not ready yet), while the
RDP_PEER_PENDING_GFX_GRAPHICS_RESET flag signals, that
gnome-remote-desktop needs to submit the monitor configuration and the
size of the graphics output buffer to the client first, in order to be
able to submit surface updates.