session-rdp: Add API to encode pending frame for RDP surface
Currently, gnome-remote-desktop encodes all pending frame data upon the
end of the SuppressOutput PDU.
When using the graphics pipeline, gnome-remote-desktop needs to be able
to control the rate of the encodings.
This means: gnome-remote-desktop needs to be able to suspend the
encoding and continue it later.
This is the case, when the client is too slow with the decoding.
In that case, gnome-remote-desktop needs to immediately stop the
encoding to not flood the client with too much new frame content.
When the client acked enough pending frames, continue the encoding
process.

To be able to do that, add an API to encode the pending frame of an RDP
surface.