rdp-surface: Add attribute to indicate that encoding is suspended
This attribute indicates, when set to TRUE, that a new frame for this
RDP surface should not be encoded yet.
It is independent of the graphics pipeline, meaning session-rdp can
later use this attribute without considering whether the graphics
pipeline is supported or not.
It will, however, later only be used for the graphics pipeline.