session-rdp: Add API to notify a severe error
The capability, that indicates the support for the graphics pipeline
for the client or the server, is already exchanged upon connection of
the RDP client.
If the client indicates support for the graphics pipeline, but actually
does not support it (opening the graphics pipeline fails), then the
client heavily violates the protocol.
In the future, gnome-remote-desktop might also require the graphics
pipeline for specific use cases, like headless sessions or RAIL, as
some actions like submitting alpha channel data (usually, when using
RAIL), submitting the monitor configuration to the client, handling
network latency can only be (easily) done with the graphics pipeline.

Additionally, if the client tries to reset the graphics pipeline with
a new capability exchange (CapsAdvertise), but is not allowed to, or
submits capability sets, that are unsupported by gnome-remote-desktop,
then gnome-remote-desktop needs to get rid of the client.

So, add an API for this use case. Normally, these severe situations
should not happen, but gnome-remote-desktop needs to be able to handle
them if they happen.