session-rdp: Add support for the graphics pipeline
Use the previously implemented graphics pipeline class to add support
for the graphics pipeline.
The graphics pipeline is a dynamic channel. It will be differently
initialized than the CLIPRDR channel, which is a static channel.
For this, the graphics pipeline needs to wait first until the DRDYNVC
channel (, which is a special static channel) is initialized, as that
channel tunnels all dynamic channels.