clipboard-rdp: Add dummy handling for the Temporary Directory PDU
In the initial implementation of the cliprdr protocol (in 1998), some
PDUs did not exist:
The capabilities PDUs, the locking PDUs and the
FileContents-Request/Response PDUs.
However, despite of that, file transfer was implemented:
Using the device redirection channel, the client advertised a drive,
where the server should put the files on a request.
The path of the redirected drive, was advertised via the Temporary
Directory PDU.
While the cliprdr documentation has this behaviour only poorly
documentated and doesn't directly mention the usage of a redirected
drive, the Windows Server 2003 and Windows XP source code leaks,
however, confirm this behaviour.
When the client requested the files, it used the FormatDataRequest with
the CF_HDROP format.
The server then copied the files to the redirected drive and sent the
file list (e.g. "temp/file1.txt temp/file2.txt") to the client.
For client-to-server transfer, the server ignored requests with the
CF_HDROP format.

Today, this PDU is unused, but clients can technically still send this
PDU.
So, add a function to the associated function pointer, so FreeRDP
doesn't dereference a NULL function pointer, upon receiving a Temporary
Directory PDU.
FileContentsRequests are used today instead, but the PDU still exists
for compatibility.

The same applies to the MonitorReady PDU, which was initially used by
the server to initiate the cliprdr protocol.
Today, the server first sends the ServerCapabilities PDU, followed by
the MonitorReady PDU to maintain backward compatibility with ancient
clients.