clipboard: Split up SelectionRead requests
In order to be able to handle SelectionRead requests in an async way,
request_server_content_for_mime_type() needs to be adjusted to not to
directly return the data.
The function now becomes a void function, meaning no data returns to
the calling function.
To return the fetched data to the clipboard implementation, use a newly
introduced vfunc.

Functionality wise, nothing changes in this commit, since the backends
still get their data in the same way, except that the calling function
is not the same any more, as the part, that processes the returned
data, is now a separate function.