session: Move 'ReadMimeTypeContent' to clipboard class
Reading the mime type content from a fd is a task for the clipboard.
Handling the read() operation in an async way should therefore be done
in grd-clipboard, as it will introduce more code, that is specific to
the clipboard.

So, move this part into grd-clipboard as another preparation for async
read() operations.
Functionality wise, nothing changes in this commit.
grd_session_selection_read() will now return the read fd, instead of
the read mime type content.