clipboard: Handle SelectionRead() operations in an async way
Currently, reading the mime type content happens by directly calling
read() on a fd.
This is problematic, as it will only work, when the other end also
supplies the mime type content.
This is not always the case and in such cases gnome-remote-desktop will
freeze with 100% cpu usage in read().
To get rid of this situation, the read() operation must happen in an
async way.
If a certain timeout passes, the read() operation needs to be aborted.

Do this using a GTask, which runs in another thread, which then runs
g_input_stream_read(), which listens on the read fd and on the
cancellable fd.
This allows g-r-d to stay responsive, even when an application does not
supply the mime type content.

When the GTask is done with the async operation, it will create a
GSource and attach it to the thread, that created the GTask.
Since that GSource function can also run, when the client is already
gone, abort the current read operation with a GCancellable, when
disposing the clipboard.

If a new mime type list is advertised by the server or client, abort
the current read operation and flush the current mime type content.
This ensures, that the order in which all clipboard operations happen
is kept, as the RDP clipboard, for example, does not define how pending
FormatDataRequests are handled, when a new FormatList is advertised.
If the read operation was successful, but the mime type content was not
submitted yet, submit the mime type content.
Otherwise, inform the backend about the abortion of the operation.
For clipboard implementations that support delayed rendering of
clipboard data (RDP), this ensures that the client is notified about
the read result.
In the case of RDP this means, that the FormatDataResponse with the
fail flag is sent.

When the read() operation is aborted, wait for the GTask thread
function to complete.
Do this using a GCond. This ensures that the read result can be
retrieved without any race conditions.
This also ensures that mutter won't deny any new SelectionRead()
requests from gnome-remote-desktop, as the fd of the pipe is closed on
gnome-remote-desktops side, before the GCond is signalled, so that
mutter won't deny the request with the "reading in parallel" error.

Fixes: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/issues/60