rdp: Introduce GFX frame log
The GFX frame log will track pending frame acks of a GFX surface and
calculate the frame encoding and frame acking rate.
The latter part will later be important, when dealing with the network
latency, as gnome-remote-desktop cannot fully rely on the measured
round trip time for handling the encoding rate, as the measured round
trip time can also reflect bottlenecks on the client side.

The GFX frame log is therefore also an initial step for network
autodetection.