rdp: Move some elements from RdpPeerContext to the GrdSessionRdp struct
In order to be able to implement separate classes, that handle Fast-
and Slowpath PDUs, the RdpPeerContext struct needs to be moved out of
the GrdSessionRdp class.
Before doing that, move some elements from the RdpPeerContext struct
into the GrdSessionRdp struct, that are private to the GrdSessionRdp
class.

This is a preparatory step for the implementation of a class, that
detects the network characteristics of the RDP session, as this class
will use and hook up to Fast- and Slowpath PDUs.