rdp-gfx-surface: Add handling to lower frame latency
While the throttling mechanism can currently take care of higher and
increasing RTT values, it can currently not take care of lowering the
latency.
For example: If an RTT value of 300ms is detected and the RDP client is
slow with the decoding and displaying process, then
gnome-remote-desktop adapts to the situation.
However, if the RTT value suddenly drops to, for example, 150ms, then
the frame content might still be delayed by 300ms, since the client
might be too slow with the frame updates.

To solve this situation, recalculate the activate threshold for the
throttling mechanism, when a new frame is encoded or a frame is being
acked.
When the new activate threshold is lower than the current one, suspend
the encoding until the client acks enough logical frames.
After this, reevaluate the throttling situation.
This ensures that gnome-remote-desktop uses the lowest possible
activate threshold for the throttling mechanism to ensure an experience
with the lowest possible latency, while still being able to handle high
latency connections.