rdp: Add class for NVENC and CUDA support
This is the first step for hardware acceleration support in
gnome-remote-desktop.
The implementation for Hardware acceleration using NVENC and CUDA is
way easier than the implementation of hardware acceleration using
VAAPI.
To be able to use NVENC and CUDA, use the ffnvcodec-headers. These
headers, provide functions to easily load NVENC and CUDA using dlopen.
NVENC itself, will be used to encode AVC420 content, which will then
be pushed via the graphics pipeline.
CUDA will be used to perform the BGRX to YUV420 colour conversion, as
that will massively improve the performance compared to colour
conversion computed on the CPU.

To be able to encode AVC420 frames, the frame size needs to be aligned
to the value of 16 (both width and height), and the colour format needs
to be converted to YUV420.
In addition to the alignment to the value of 16, NVENC seems to require
the frame height to be aligned to a multiple of 64, as otherwise the
resulting frame on the client may contain a black strip in the middle
of the frame, when the height is not aligned to a multiple of 64.
Since the FreeRDP primitives are way too slow (17-24ms for a FullHD
frame) to take care of the colour conversion, use a CUDA kernel to
perform this operation.
This will reduce the conversion time under 400µs (313µs according to
the NVIDIA Visual Profiler) on a GeForce GTX 660.

The resulting image (YUV420 in the NV12 format) will then be passed to
NVENC, which then encodes the frame.
When using NVENC with MBAFF, NVENV requires the image to be already
interlaced.
If it is not interlaced, then even lines end up in the resulting image
at the position y / 2, instead of y, while odd lines end up in the
resulting image at the position y / 2 + aligned_height / 2, instead of
y.
To take care of this situation, calculate the interlaced position
directly in the CUDA kernel function.
The resulting image will then be correct on the client side.

NVENC support was introduced with the Kepler generation.
Since the CUDA toolkit removed Kepler support with version 11, and most
distributions don't ship a CUDA package, ship the generated PTX code
with gnome-remote-desktop.
The PTX code is generated with the CUDA toolkit version 10 and will
work for all Kepler and later GPUs, since PTX code is forward
compatible.

PTX code is not binary, meaning it is a human readable text file.
CUDA code is generated in two steps:

First, the PTX code: The PTX code is code, that is generated for a
specific compute capability, but can also be processed by GPUs, that
support a newer compute capability.
However, it cannot be processed by GPUs with an older (lower) compute
capability.

Second, the CUDA binary: The binary that will end up on the GPU
eventually.
Technically, gnome-remote-desktop could ship that binary, but this is
not suitable:
First, it is a binary, it cannot be easily verified.
Second, the binary is GPU specific, meaning gnome-remote-desktop would
have to ship a fat binary to cover all GPUs, which is not suitable.

The NVIDIA driver ships a JIT compiler, which can load PTX code and
generate the CUDA binary at runtime.

When gnome-remote-desktop starts, the NVIDIA driver will automatically
use the JIT compiler to produce the device specific CUDA binary.
This is a fast process and it will also load the module.

gnome-remote-desktop then uses the module to perform the colour
conversion, when encoding an AVC420 frame.

In the future, the NVENC and CUDA implementation will be extended to be
able to produce AVC444 frames.
The NVENC capable GPU doesn't need to have support for AVC444 frames
for this, since RDP uses a special way to create AVC444 frames
(composed out of two AVC420 frames, one main view, one auxiliary view).