rdp: Add mechanism to lower or increase the ping interval
With the API to lower or increase the ping interval in place, add a
mechanism that uses the previously implemented API to change the ping
interval depending on the encoding rate (and therefore for the need for
new and fresh round trip times).

For the mechanism, add a new GSource, which checks every second the
amount of surface updates.
If the amount reaches zero, destroy the GSource and notify the network
autodetection class to lower the ping interval.
If the encoding starts again attach the (recreated) GSource again and
notify the network autodetection class to increase the ping interval.