rdp/nw-auto: Add API to allow lowering or increasing the ping interval
When no frame updates happen for some time, but the client window is
not minimized, gnome-remote-desktop will currently happily still emit
a lot of RTTRequests, producing unnecessary network activity.
This usually results in a bandwidth usage of ~2.5KiB/s.

When the graphics pipeline detects that the global encoding rate
stalled at zero, it should be able to notify the network autodetection
class about it, so that the network autodetection class can lower the
ping interval.
Lowering the ping interval, instead of stopping the detection mechanism
completely, still allows gnome-remote-desktop to track the current
round trip time to ensure that the throttling mechanism still has an up
to date value of the round trip time, when gnome-remote-desktop
continues to encode frame content.
When any encoding activity happens again, the graphics pipeline will
notify the network autodetection class again about the behaviour, to
increase the ping interval again for a more accurate round trip time.