clipboard-rdp: Fix comments and output messages
CLIPRDR does not perform stream file clipping, it performs streaming
operations of file clips.

So, fix that output message. Also fix some comments, to use the word
"notify", instead of "inform", and fix a comment to correctly reflect
the documentation ("can now be released" -> "MUST now be released").
Additionally, set the copyright year correctly, as the clipboard data locking
implementation was done in 2021.