clipboard-rdp: Don't leak memory when session ends
When the remote desktop session ends, but there is still a list of
pending mime type tables for the server, then these tables are leaked.
While it is in reality very unlikely that this situation happens, it is
theoretically not impossible that this situation happens.

So, save the FormatListUpdateContext, like the FormatDataRequestContext
on the clipboard_rdp struct to allow freeing the memory, when the
associated server_format_list_update_id is cleared.