clipboard-rdp: Don't add duplicated mime types to mime type tables
Normally, all FormatLists contain all announced formats only once.
With xfreerdp3, this is not the case any more, since the introduction
of server to client file transfer via the clipboard for xfreerdp.
Since file lists are announced via the name "FileGroupDescriptorW" and
their id is dynamically assigned, gnome-remote-desktop replaces the
existing mime type tables of their mime type.

The problem with duplicated mime types is, that gnome-remote-desktop
correctly replaces the old mime type tables, if they exist, but for
each freed mime type table still tries to announce them to mutter.
glib realizes this situation, when creating the string variant for each
type and emits a critical warning (instead of crashing).

To solve this situation, use a GHashTable to ensure that mime type
tables are only added once.
Once all mime type tables are in the list, destroy the temporary hash
table.
With this handling, gnome-remote-desktop ignores all duplicated entries
in a FormatList and only picks their first occurrence.