rdp: Add classes for audio output forwarding
In order to handle audio output redirection and forwarding,
gnome-remote-desktop needs to be able to do the following:

1. Keep track of the audio sinks
2. Know, when audio is actually playing on an audio sink
3. Know the current volume and muted state of a sink
4. Keep track of the render latency of the client
5. Buffer the audio data up and encode that buffer
6. Drop old sample data, when buffered data gets too old

For task 1, add callbacks for the global() and global_remove events of
the PipeWire registry.
PipeWire will notify gnome-remote-desktop with these events, when an
object is made available to PipeWire or removed.
Use these events to find the audio sinks by examining the object
properties, when such global event is emitted.

When an audio sink is found, create a class instance for the new audio
output stream for that node.
This instance will keep track of that node to listen for parameter
changes such as volume changes and whether the audio was muted on that
sink (task 3).
This instance will also create a PipeWire stream for the given node.
In order to be able to maintain a stable render latency on the client
side, force a quantum of 256.
With this quantum, PipeWire submits every 5-6ms new audio data.

When connecting a PipeWire stream to an audio sink, the stream consumer
will always receive audio data, even when no audio is playing.
Additionally, it is not known via PipeWire, which audio sink is the
primary one, so gnome-remote-desktop has to find this out (task 2).

To handle this situation, use a lock()/unlock() method in combination
with methods, that determines, whether the audio data is empty (no
amplitude of the signals) or whether the volume is muted.

When audio plays, lock that sink in. The stream for that sink will then
have the privilege to be allowed to submit audio data to the pending
frames queue.
When no more audio data is played on that sink, unlock the respective
stream, but don't do it immediately, as it would cause stutters:

When sending samples to the client, the client usually first buffers
the audio data up on the sink on its side.
When there is a short pause in the audio data, e.g. due to a pause in a
song, the audio will stutter, when continuing the audio stream.
As a result, empty audio buffers should here be sent to the client to
take care of this situation.

However, empty audio buffers should not be always sent to the client:
First, it's a waste of bandwidth and second, there may also be other
audio sinks on the server side.
The user might have switched the audio sink and gnome-remote-desktop
needs to adapt to the situation.

As a result, set a timestamp, when receiving an empty buffer.
When more empty buffers are received and a timeout is reached, unlock
the stream.
When the timeout has not been reached yet and a non-empty buffer is
received, unset the state, that a timestamp was set.

When locking a stream, set the other streams as inactive.
This basically pauses the other streams, to reduce the usage of system
resources.
When unlocking a stream, set all streams as active.
Since setting a stream as inactive is a blocking operation, where
PipeWire waits for the real time thread to finish its task, do not start
any streams in the active state.
Since acquiring and releasing the stream lock happens, while a mutex is
locked, this would otherwise lead to a deadlock.
Starting a stream as inactive and only activating it, when no stream
lock is set, gets rid of this issue.

When sending encoded samples as a block, the client will normally
confirm this block two times:
First, upon retrieval and second, when the audio data was played.
However, there is no guarantee, that a client sends both Wave Confirm
PDUs.
This will especially be the case in the future, when transmitting the
audio data via UDP.

To take care of this situation, do not submit a timestamp for each sent
block (sent timestamp == 0).
When receiving a Wave Confirm PDU, only update the rendered latency,
when the timestamp is non-zero, as that timestamp will then represent
the rendered latency.
When the rendered latency gets too high, drop samples to get rid of the
async audio (task 4).

In addition to that, also set a timestamp for each submitted audio
buffer from a stream, that is submitted to the pending frames queue.
When situations, like a system hang happen, the audio may become async
too.
When the buffer age gets too old, drop these samples (task 5).

When encoding audio data, not every buffer size matches the
requirements of certain codecs.
This is especially the case with fdk-aac.
Normally, PipeWire streams submit a buffer size of 4096.
However, when playing different audio, this buffer size may be smaller
or bigger depending on the producer of the audio.
In addition to that, the quantum will be different.
PipeWire, by-default, always use the lowest quantum of all streams for
an audio sink, but individual streams can have different (forced)
quantums.
Also, mstsc does not handle audio streams with more than 44100 samples
per second.
Audio streams with more than 44100 samples per second are treated as
audio streams with 44100 samples per second by mstsc.
As a result, the audio streams need to be resampled, which leads to
different (usually smaller) buffer sizes, that are submitted to
gnome-remote-desktop.

To take care of this situation, put all submitted audio buffers in a
queue (pending frames queue).
When encoding the buffer, take the needed pieces of buffers from that
queue (task 5).
To reduce the additional latency, use a quantum of 256 for the PipeWire
streams (audio buffer every 5-6ms).
This is also the same quantum, that is used, when running most games.